CH: restructure file docstrings
