+++ src/services/auth.ts
@@ -1,12 +1,19 @@
+/** 
+ * @file auth.ts
+ * 
+ * @description
+ * This file contains everything that requires accessing Firebase Authentication
+ * 
+ * @exports signInWithGoogle - Function opening a popup allowing the user to sign in with Google
+ * @exports handleLogout - Function logging the user out from auth
+ */ 
+
+
 // External Libraries
 import { signInWithPopup, GoogleAuthProvider, signOut } from "firebase/auth";
 
 // Internal Modules
 import { auth } from "./firebase"
-
-/** 
- * This file contains everything that requires accessing Firebase Authentication 
- */ 
 
 
 // Sets up google provider
