+++ src/services/storage.ts
@@ -1,3 +1,13 @@
+/** 
+ * @file storage.ts
+ * 
+ * @description
+ * This module contains everything that requires accessing Firebase Storage
+ * 
+ * @exports saveProfilePic - Saves profile picture to storage
+ * @exports getProfilePic - Gets the profile pic from storage
+ */ 
+
 // External Libraries
 import { ref, uploadString, getDownloadURL } from "firebase/storage";
 
@@ -5,13 +15,9 @@ import { ref, uploadString, getDownloadURL } from "firebase/storage";
 import { storage } from "./firebase";
 
 
-/** 
- * This module contains everything that requires accessing Firebase Storage
- */ 
-
 
 /**
- * Saves profile picture to database
+ * Saves profile picture to storage
  * @param uid - The unique id for the user
  * @param profilePicURL - The profile picture stored as a URL
  */
@@ -21,7 +27,7 @@ export function saveProfilePic(uid: string, profilePicURL: string) {
 }
 
 /**
- * Gets the profile pic from stoarage
+ * Gets the profile pic from storage
  * @param location - The location of the profile pic in storage
  * @returns The download URL
  */
