+++ src/types/aliases.ts
@@ -1,6 +1,7 @@
 /**
- * aliases.ts
+ * @file aliases.ts
  * 
+ * @description
  * This file contains any type aliases used.
  * These aliases definine complex or reusable types with a shorter, more descriptive name.
  */
