+++ src/utils/text-formatting.tsx
@@ -1,13 +1,19 @@
+/** 
+ * @file text-formatting.tsx
+ * 
+ * @description
+ * Utility functions related to the formatting and styling of text
+ * This includes markdown and LaTeX
+ * 
+ * @exports markdownToHTML - A function converting markdown and LaTeX formatting to HTML
+ */ 
+
+
 // Import External Libraries
 import { ReactElement } from 'react';
 import ReactMarkdown from 'react-markdown';
 import 'katex/dist/katex.min.css';
 import Latex from 'react-latex-next';
-
-/** 
- * @file This module contains everything that to do with formatting text 
- * @module TextFormatting
- */ 
 
 /**
  * Converts a sting into a series of markdown React components
